+++ hardpy/hardpy_panel/frontend/src/hardpy_test_view/TestSuite_v3.tsx
@@ -17,7 +17,7 @@ import RunTimer from './RunTimer';
 import './TestSuite.css';
 import { Spin } from 'antd';
 
-interface WidgetInfo{
+interface WidgetInfo {
     info: Record<string, unknown>;
     type: WidgetType;
 }
@@ -26,7 +26,7 @@ interface DialogBoxProps {
     title_bar?: string;
     dialog_text: string;
     widget?: WidgetInfo
-  }
+}
 
 interface Case {
     status: string,
@@ -86,16 +86,16 @@ export class TestSuite extends React.Component<Props, State> {
                             <div style={{ display: 'flex', alignItems: 'center' }}>
                                 <TestStatus
                                     status={(
-                                            this.props.commonTestTunStatus != 'run'
-                                            && (this.props.test.status == 'run' || this.props.test.status == 'ready')
-                                        ) ? "stucked" : this.props.test.status
+                                        this.props.commonTestTunStatus != 'run'
+                                        && (this.props.test.status == 'run' || this.props.test.status == 'ready')
+                                    ) ? "stucked" : this.props.test.status
                                     }
                                 />
                                 <Icon
                                     style={{ marginRight: '10px', marginLeft: '10px' }}
                                     icon={this.state.isOpen ? 'chevron-down' : 'chevron-right'}
                                 ></Icon>
-                                <span>{this.renderName(this.props.test.name, this.props.index+1)}</span>
+                                <span>{this.renderName(this.props.test.name, this.props.index + 1)}</span>
                             </div>
                         </Button>
                     </div>
@@ -128,16 +128,15 @@ export class TestSuite extends React.Component<Props, State> {
         return (
             <H4 className={`test-suite-name ${is_loading ? Classes.SKELETON : ""}`}>
                 {<span className={Classes.TEXT_DISABLED}>{test_number}</span>}
-                {<span style={{marginLeft: '0.5em'}}>{is_loading ? SUITE_NAME_STUB : name}</span>}
+                {<span style={{ marginLeft: '0.5em' }}>{is_loading ? SUITE_NAME_STUB : name}</span>}
             </H4>
         );
     }
 
     private renderTests(test_topics: Cases) {
-        let case_names : string[] = []
+        let case_names: string[] = []
 
-        if (test_topics)
-        {
+        if (test_topics) {
             case_names = Object.keys(test_topics)
         }
 
@@ -207,13 +206,13 @@ export class TestSuite extends React.Component<Props, State> {
 
                 <Tag minimal={true} style={{ margin: '2px', minWidth: '30px' }}>
                     {
-                    ('ready' != test_topics.status) &&
-                    (
-                        <RunTimer
-                            status={test_topics.status}
-                            commonTestTunStatus={this.props.commonTestTunStatus}
-                        />
-                    )}
+                        ('ready' != test_topics.status) &&
+                        (
+                            <RunTimer
+                                status={test_topics.status}
+                                commonTestTunStatus={this.props.commonTestTunStatus}
+                            />
+                        )}
                 </Tag>
             </div>
         );
@@ -234,7 +233,7 @@ export class TestSuite extends React.Component<Props, State> {
     private cellRendererNumber(test_topics: Case[], row_: string, rowIndex: number) {
         return this.commonCellRender(
             <div style={{ marginTop: '0.2em', marginBottom: '0.2em' }}>
-                <TestNumber val={rowIndex+1} />
+                <TestNumber val={rowIndex + 1} />
             </div>,
             `number_${rowIndex}_${row_}}`,
         );
@@ -246,7 +245,7 @@ export class TestSuite extends React.Component<Props, State> {
             <div style={{ marginTop: '0.2em', marginBottom: '0.2em' }}>
                 <TestName name={test.name} />
             </div>,
-             `name_${rowIndex}_${row_}`
+            `name_${rowIndex}_${row_}`
         );
     }
 
@@ -266,27 +265,27 @@ export class TestSuite extends React.Component<Props, State> {
         const { info: widget_info, type: widget_type } = test.dialog_box.widget || {};
 
         return this.commonCellRender(
-          <div style={{ marginTop: '0.2em', marginBottom: '0.2em' }}>
-            {test.dialog_box.dialog_text && (
-              <StartConfirmationDialog 
-                title_bar={test.dialog_box.title_bar || test.name}
-                dialog_text={test.dialog_box.dialog_text}
-                widget_info={widget_info}
-                widget_type={widget_type}
+            <div style={{ marginTop: '0.2em', marginBottom: '0.2em' }}>
+                {test.dialog_box.dialog_text && (
+                    <StartConfirmationDialog
+                        title_bar={test.dialog_box.title_bar || test.name}
+                        dialog_text={test.dialog_box.dialog_text}
+                        widget_info={widget_info}
+                        widget_type={widget_type}
+                    />
+                )}
+                <TestStatus
+                    status={
+                        (this.props.commonTestTunStatus !== 'run' &&
+                            (test.status === 'run' || test.status === 'ready'))
+                            ? "stucked"
+                            : test.status
+                    }
                 />
-            )}
-            <TestStatus
-              status={
-                (this.props.commonTestTunStatus !== 'run' &&
-                  (test.status === 'run' || test.status === 'ready'))
-                  ? "stucked"
-                  : test.status
-              }
-            />
-          </div>,
-          `status_${rowIndex}_${row_}`
+            </div>,
+            `status_${rowIndex}_${row_}`
         );
-      }
+    }
 
     private handleClick = () => this.setState((state: State) => ({ isOpen: !state.isOpen }));
 }