+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+* Add dialog boxes.
 * Add a socket mechanism to transfer data from the uvicorn server to the pytest subprocess.
 
 ## HardPy 0.3.0