+++ docs/changelog.md
@@ -3,6 +3,9 @@
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
 * Add dialog box with radiobutton and checkbox.
+
+## HardPy 0.4.0
+
 * Add base dialog box, text input and numeric input.
 * Add dialog box invocation functionality.
 * Add a socket mechanism to transfer data from the uvicorn server to the pytest subprocess.